feat(settings): unify athlete profile and account settings under /settings
Consolidate /compete/athlete/* settings routes into a single /settings/*
namespace with a grouped sidebar (General / Athlete / Account). Adds
Overview and Appearance tabs and 301-style redirects from old paths.

- Sidebar: grouped nav (Overview, Profile, Appearance | Athlete profile,
  Teams | Security, Sessions, Billing) with Programming nested under
  Athlete when workout tracking is on
- /settings/overview: identity strip + profile-completion checklist that
  deep-links into the right tab for each unfinished item
- /settings/athlete, /sponsors, /billing: same forms moved off
  /compete/athlete/{edit,sponsors,invoices}; old paths now redirect
- /settings/appearance: light/dark theme picker (persists in
  localStorage)
- /settings index now redirects to /settings/overview
- Internal links and post-login redirects updated to point at the new
  /settings/* paths

Tests cover the new sidebar groups + nav items and the
calculateProfileCompletion helper.
