+++ apps/wodsmith-start/src/routes/_protected/settings/index.tsx
@@ -3,7 +3,7 @@ import { createFileRoute, redirect } from "@tanstack/react-router"
 export const Route = createFileRoute("/_protected/settings/")({
   beforeLoad: () => {
     throw redirect({
-      to: "/settings/profile",
+      to: "/settings/overview",
     })
   },
 })
