+++ apps/wodsmith-start/src/routes/compete/athlete/invoices/index.tsx
@@ -1,161 +1,7 @@
-import { createFileRoute, Link } from "@tanstack/react-router"
-import { ArrowLeft, ChevronRight, FileText, Receipt } from "lucide-react"
-import { Badge } from "@/components/ui/badge"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent } from "@/components/ui/card"
-import { getAthleteInvoicesDataFn } from "@/server-fns/athlete-profile-fns"
+import { createFileRoute, redirect } from "@tanstack/react-router"
 
 export const Route = createFileRoute("/compete/athlete/invoices/")({
-  component: InvoicesPage,
-  loader: async () => {
-    return await getAthleteInvoicesDataFn()
+  beforeLoad: () => {
+    throw redirect({ to: "/settings/billing" })
   },
 })
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-function formatCurrency(cents: number): string {
-  return new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency: "USD",
-  }).format(cents / 100)
-}
-
-function formatDate(date: Date | null): string {
-  if (!date) return "-"
-  return new Intl.DateTimeFormat("en-US", {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-  }).format(new Date(date))
-}
-
-function getStatusBadge(status: string) {
-  switch (status) {
-    case "COMPLETED":
-      return <Badge variant="default">Paid</Badge>
-    case "PENDING":
-      return <Badge variant="secondary">Pending</Badge>
-    case "FAILED":
-      return <Badge variant="destructive">Failed</Badge>
-    case "CANCELLED":
-      return <Badge variant="outline">Cancelled</Badge>
-    default:
-      return <Badge variant="outline">{status}</Badge>
-  }
-}
-
-// ============================================================================
-// Component
-// ============================================================================
-
-/**
- * Group purchases by checkout session so multi-division registrations
- * show as a single invoice row.
- */
-function groupByInvoice(
-  purchases: ReturnType<typeof Route.useLoaderData>["purchases"],
-) {
-  const groups = new Map<
-    string,
-    { primary: (typeof purchases)[number]; totalCents: number }
-  >()
-  const standalone: Array<{
-    primary: (typeof purchases)[number]
-    totalCents: number
-  }> = []
-
-  for (const p of purchases) {
-    const key = p.stripeCheckoutSessionId
-    if (!key) {
-      standalone.push({ primary: p, totalCents: p.totalCents })
-      continue
-    }
-    const existing = groups.get(key)
-    if (existing) {
-      existing.totalCents += p.totalCents
-    } else {
-      groups.set(key, { primary: p, totalCents: p.totalCents })
-    }
-  }
-
-  return [...groups.values(), ...standalone]
-}
-
-function InvoicesPage() {
-  const { purchases } = Route.useLoaderData()
-  const invoices = groupByInvoice(purchases)
-
-  return (
-    <div className="mx-auto max-w-4xl space-y-6 pb-12">
-      {/* Header */}
-      <div className="flex items-center gap-4">
-        <Button variant="ghost" size="icon" asChild>
-          <Link to="/compete/athlete">
-            <ArrowLeft className="h-4 w-4" />
-          </Link>
-        </Button>
-        <div className="flex items-center gap-2">
-          <Receipt className="h-6 w-6" />
-          <h1 className="font-bold text-2xl">Invoices</h1>
-        </div>
-      </div>
-
-      {invoices.length === 0 ? (
-        <Card>
-          <CardContent className="py-12 text-center">
-            <FileText className="mx-auto h-12 w-12 text-muted-foreground/50" />
-            <p className="mt-4 text-muted-foreground">No invoices yet</p>
-            <p className="text-muted-foreground text-sm">
-              When you register for competitions, your invoices will appear
-              here.
-            </p>
-          </CardContent>
-        </Card>
-      ) : (
-        <div className="space-y-3">
-          {invoices.map(({ primary: purchase, totalCents }) => (
-            <Link
-              key={purchase.id}
-              to="/compete/athlete/invoices/$purchaseId"
-              params={{ purchaseId: purchase.id }}
-            >
-              <Card className="transition-colors hover:bg-muted/50">
-                <CardContent className="flex items-center justify-between py-4">
-                  <div className="flex items-center gap-4">
-                    <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-primary/10">
-                      <Receipt className="h-5 w-5 text-primary" />
-                    </div>
-                    <div>
-                      <p className="font-medium">
-                        {purchase.competition?.name ?? purchase.product.name}
-                      </p>
-                      <p className="text-muted-foreground text-sm">
-                        {formatDate(purchase.completedAt ?? purchase.createdAt)}
-                        {purchase.competition?.organizingTeam && (
-                          <span>
-                            {" "}
-                            &middot; {purchase.competition.organizingTeam.name}
-                          </span>
-                        )}
-                      </p>
-                    </div>
-                  </div>
-                  <div className="flex items-center gap-4">
-                    {getStatusBadge(purchase.status)}
-                    <span className="font-medium">
-                      {formatCurrency(totalCents)}
-                    </span>
-                    <ChevronRight className="h-4 w-4 text-muted-foreground" />
-                  </div>
-                </CardContent>
-              </Card>
-            </Link>
-          ))}
-        </div>
-      )}
-    </div>
-  )
-}
